Moving shipping classes from `woocommerce/woocommerce-services/state` to `woocommerce/state` because shipping classes belong to WooCommerce and not WooCommerce Services.
