+++ client/extensions/woocommerce/state/data-layer/shipping-classes/index.js
@@ -8,7 +8,7 @@ import request from 'woocommerce/state/sites/http-request';
 import {
 	fetchShippingClassesFailure,
 	fetchShippingClassesSuccess,
-} from 'woocommerce/woocommerce-services/state/shipping-classes/actions';
+} from 'woocommerce/state/shipping-classes/actions';
 import { WOOCOMMERCE_SERVICES_SHIPPING_CLASSES_REQUEST } from 'woocommerce/woocommerce-services/state/action-types';
 import { verifyResponseHasData } from 'woocommerce/state/data-layer/utils';
