+++ client/extensions/woocommerce/state/shipping-classes/reducers.js
