+++ client/extensions/woocommerce/state/shipping-classes/selectors.js
