+++ client/extensions/woocommerce/state/shipping-classes/test/data/initial-state.js
