+++ client/extensions/woocommerce/state/shipping-classes/test/reducers.js
