+++ client/extensions/woocommerce/state/shipping-classes/test/selectors.js
